Fri, Jul 28, 2023, 4:17 PM +06:00

+++ beecrowd/1255.c
@@ -18,8 +18,8 @@ int main(void)
         int j = 0;
         while(text[j] != '\0')
         {
-            if((text[j] >= 65) && (text[j] <= 90)) freq[(int)text[j] + 32]++;
-            else freq[(int)text[j]]++;
+            if((text[j] >= 65) && (text[j] <= 90)) freq[text[j] + 32]++;
+            else freq[text[j]]++;
             j++;
         }
